cclient/serialization/json: split up tests
