+++ cclient/serialization/BUILD
@@ -28,15 +28,3 @@ cc_library(
         "//cclient/response:responses",
     ],
 )
-
-cc_test(
-    name = "test_json",
-    srcs = [
-        "tests/test_json.c",
-        "tests/test_json.h",
-    ],
-    deps = [
-        ":serializer_json",
-        "@unity",
-    ],
-)
