+++ cclient/serialization/json/tests/BUILD
@@ -1,0 +1,144 @@
+cc_library(
+    name = "shared",
+    hdrs = ["shared.h"],
+)
+
+cc_test(
+    name = "add_neighbors",
+    srcs = ["add_neighbors.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "attach_to_tangle",
+    srcs = ["attach_to_tangle.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "broadcast_transactions",
+    srcs = ["broadcast_transactions.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "check_consistency",
+    srcs = ["check_consistency.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "find_transactions",
+    srcs = ["find_transactions.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "get_balances",
+    srcs = ["get_balances.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "get_inclusion_states",
+    srcs = ["get_inclusion_states.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "get_neighbors",
+    srcs = ["get_neighbors.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "get_node_info",
+    srcs = ["get_node_info.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "get_tips",
+    srcs = ["get_tips.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "get_transactions_to_approve",
+    srcs = ["get_transactions_to_approve.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "get_trytes",
+    srcs = ["get_trytes.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "remove_neighbors",
+    srcs = ["remove_neighbors.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
+
+cc_test(
+    name = "store_transactions",
+    srcs = ["store_transactions.c"],
+    deps = [
+        ":shared",
+        "//cclient/serialization:serializer_json",
+        "@unity",
+    ],
+)
